加班費改丟了,

+++ app/src/main/java/com/ray/trarailwaysalaryapp/EmployeeSalary.kt
@@ -1,9 +1,8 @@
 package com.ray.trarailwaysalaryapp
 
-// 從業人員的薪資數據模型
-// IMPORTANT: This data class should ONLY contain 'grade' and 'amount'.
-// The 'professionalAllowance' is now retrieved from the EmployeePosition enum.
+import java.math.BigDecimal // 確保有導入 BigDecimal
+
 data class EmployeeSalary(
-    val grade: Int,             // 薪點 (B欄)
-    val amount: Double          // 薪額 (C欄)
+    val grade: Int,
+    val amount: BigDecimal // <-- 這裡必須是 BigDecimal
 )
