+++ app/src/main/java/com/ray/trarailwaysalaryapp/OperatorSalary.kt
@@ -1,8 +1,8 @@
 package com.ray.trarailwaysalaryapp
 
-// 營運人員的薪資數據模型
+import java.math.BigDecimal // 確保有導入 BigDecimal
+
 data class OperatorSalary(
-    val grade: Int,             // 薪點 (D欄)
-    val amount: Double          // 月支數額 (E欄)
-    // val talentRetentionAllowance: Double? = null // 移除此行，留才津貼現在是固定值
+    val grade: Int,
+    val amount: BigDecimal // <-- 這裡必須是 BigDecimal
 )
